Use getBlockInventory to get chest inventories
The getBlockInventory method returns the inventory of an individual chest block. If the block is part of a double chest, it returns that block's half of the double chest inventory. Using this method makes it possible to pull materials from double chests since their full inventories are no longer accessed twice.